+++ src/components/Editor/Runners/PythonRunner/PyodideRunner/VisualOutputPane.jsx
@@ -59,6 +59,15 @@ const showVisual = (visual, output) => {
     case "turtle":
       output.current.innerHTML = elementFromProps(visual.content).outerHTML;
       break;
+    case "matplotlib":
+      // convert visual.content from Uint8Array to jpg
+      const img = document.createElement("img");
+      img.style = "max-width: 100%; max-height: 100%;";
+      img.src = `data:image/jpg;base64,${window.btoa(
+        String.fromCharCode(...new Uint8Array(visual.content)),
+      )}`;
+      output.current.innerHTML = img.outerHTML;
+      break;
     default:
       throw new Error(`Unsupported origin: ${visual.origin}`);
   }