+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Enabling `pyodide` support in the web component (#1090)
+- `Pyodide` `matplotlib` support (#1087)
 
 ## [0.27.1] - 2024-10-01